bot/requirements.txt
aiogram==2.9.2
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.10
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
multidict==5.2.0
psycopg2-binary==2.9.3
pytz==2021.3
requests==2.27.1
SQLAlchemy==1.4.29
urllib3==1.26.8
yarl==1.7.2
